Реализованы сервисы и добавлены необходимые мапперы

+++ src/main/java/com/doomsday/tournamentserver/service/PrizePlaceService.java
@@ -6,9 +6,10 @@ import com.doomsday.tournamentserver.service.model.view.PrizePlaceView;
 import java.util.List;
 
 public interface PrizePlaceService {
+
     boolean update(long idUser, PrizePlaceView prizePlace);
     PrizePlace getPrizePlace(long idUser, long id);
-    PrizePlace getPrizePlace(long idUser, long idTournament, long number);
+    PrizePlace getPrizePlace(long idUser, long idTournament, int number);
     List<PrizePlace> getPrizePlaces(long idUser, long idTournament);
     List<PrizePlace> getPrizePlaces(long idUser, long idTournament, boolean state);
 }
